requirements.txt
scipy
scikit-learn
numpy
opencv
Pillow
pytorch
torchvision
